requirements.txt
python-dotenv
pandas
numpy
faiss
os
json
re
google.generativeai 
requests
dotenv